chore(arcane/physics): length-literal audit for MKS (P1 stage iii sweep)
Systematic sweep of Arcane/Core/src/Arcane/Physics/** for numeric literals
carrying LENGTH or VELOCITY semantics that were tuned for pixel scale. One
straggler fixed; SpatialGrid guard-bound comments restated for MKS. Every hit
verdicted {unit-free | already-MKS | fixed | deferred-to-phase-N}; full table
in .superpowers/sdd/task-4-report.md.

Fixes (both byte-identical -- dormant default + comment-only, no re-baseline):
- SpatialHash ctor default cellSize 64 -> 1.0 m (SpatialHash.hpp:46,83). The
  live path (MakeBroadphase) always passes WorldDef::hashCellSize (=1.0 after
  T2) and every test passes an explicit cell size; a repo-wide grep found ZERO
  argless SpatialHash() -- the default was dormant everywhere, but a latent
  64 m landmine. Aligns it with WorldDef::hashCellSize and spec s3; BucketKey
  bound comment restated.
- SpatialGrid kMaxCellsPerAxis / kMaxCellsTotal comments -> MKS reading
  (65536 cells x 1 m ~= 65 km per-axis bound; 1<<22 ~= 2 km x 2 km budget).
  Values unchanged: the guards are cell COUNTS (scale-relative), not lengths.

Verdict highlights:
- already-MKS: WorldDef defaults (grav +10, sleep .05, restitution 1, push 3,
  maxLinVel 400), kSkin .02, kLinearSlop .005, DynamicTree::kMargin .05,
  residency tile 1.0, pad floor = kMargin (T3), threshSq = (kSkin/dt)^2.
- unit-free: kBulletEpsilon .001 (a sweep FRACTION back-off in [0,1], not a
  length -- refines spec's "meter-sane"), 1e-6/1e-9/1e-12 near-zero guards,
  0.94/0.342 angle/dot thresholds, 0.5 midpoints, 32/64 iteration+vertex caps,
  32-byte SIMD alignment, mass/inertia divisors, hertz/damping/beta ratios.
- deferred (cluster-confined; flipping now breaks px-scale cluster tests, CC
  precedent): Baumgarte kSlop 0.5 (length) + kRestVel 20 (velocity) -> P2
  (test-selected PGS oracle; live SoftStep reads thresholds from WorldDef);
  Gjk kShapeCastTol 0.05 (length; Box2D derives from B2_LINEAR_SLOP,
  distance.c:611-614) -> P4/P5 (Queries/CCD); Joint maxForce 1e9 (force,
  scale-coupled content) -> P5; CharacterController retunes -> P4 (out of scope).

Suite GREEN, byte-identical: ~[gpu] 494 cases / 110295 asserts, reconciled from
[physics] 279/28881 + ~[gpu]~[physics] 217/81426 - [physics][gpu] 2/12.

+++ Core/src/Arcane/Physics/Broadphase/SpatialHash.hpp
@@ -6,7 +6,7 @@
 // broadphase behind IBroadphase (the DEFAULT is DynamicTree).
 //
 // Faithful ports of the Lua algorithm:
-//   * Uniform grid of cellSize (default 64). Buckets keyed kx*0x40000 + ky.
+//   * Uniform grid of cellSize (default 1.0 m, MKS). Buckets keyed kx*0x40000 + ky.
 //   * Per-id range tracking {kx0,ky0,kx1,ky1}; update skips re-bucketing when
 //     the cell range is unchanged (the Lua's update-skip-if-same-range).
 //   * remove clears the id from its bucket range exactly.
@@ -42,8 +42,13 @@ namespace Arcane
         class SpatialHash final : public IBroadphase
         {
         public:
-            // cellSize default 64 (SpatialHash.new(cellSize or 64)).
-            explicit SpatialHash(Real cellSize = Real(64));
+            // cellSize default 1.0 m (MKS). Mirrors WorldDef::hashCellSize (the
+            // value the world actually passes via MakeBroadphase); this argless
+            // default is dormant on every live/test path but is aligned so a
+            // future argless construction cannot silently mean a 64 m cell. Was
+            // 64 px, ported from SpatialHash.new(cellSize or 64); flipped for MKS
+            // units (spec 2026-07-02 s3, matching WorldDef::hashCellSize -> 1.0).
+            explicit SpatialHash(Real cellSize = Real(1));
 
             void Update(std::uint32_t id, const Aabb2& box) override;
             void Remove(std::uint32_t id) override;
@@ -66,8 +71,8 @@ namespace Arcane
             };
 
             // Bucket key (the Lua bkey: kx*0x40000 + ky).
-            // Collision-free while |ky| < 0x20000 (131072 cells on Y; at
-            // cellSize=64 that is +/-8.4M world units).
+            // Collision-free while |ky| < 0x20000 (131072 cells on Y; at the
+            // 1.0 m default cell that is +/-131072 m ~= +/-131 km).
             [[nodiscard]] static std::int64_t BucketKey(std::int32_t kx,
                                                         std::int32_t ky) noexcept
             {
@@ -80,7 +85,7 @@ namespace Arcane
 
             void RemoveFromBuckets(std::uint32_t id, const CellRange& r);
 
-            Real m_cellSize = Real(64);
+            Real m_cellSize = Real(1);  // MKS default; see the ctor comment.
 
             // bucket key -> ids in that cell. unordered_map iteration order is
             // NOT used for output (pairs() iterates ids ascending via m_entries
